+++ src/main/java/de/maxhenkel/corpse/gui/ITransferrable.java
@@ -1,0 +1,7 @@
+package de.maxhenkel.corpse.gui;
+
+public interface ITransferrable {
+
+    void transferItems();
+
+}