+++ src/main/java/de/maxhenkel/corpse/net/MessageShowCorpseInventory.java
@@ -1,16 +1,12 @@
 package de.maxhenkel.corpse.net;
 
 import de.maxhenkel.corelib.net.Message;
-import de.maxhenkel.corpse.Death;
-import de.maxhenkel.corpse.DeathManager;
-import de.maxhenkel.corpse.entities.CorpseEntity;
-import de.maxhenkel.corpse.gui.CorpseContainerProvider;
+import de.maxhenkel.corpse.gui.Guis;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.network.PacketBuffer;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.fml.network.NetworkEvent;
-import net.minecraftforge.fml.network.NetworkHooks;
 
 import java.util.UUID;
 
@@ -38,20 +34,8 @@ public class MessageShowCorpseInventory implements Message {
         PlayerEntity player = context.getSender().world.getPlayerByUuid(playerUUID);
 
         if (player instanceof ServerPlayerEntity) {
-            openCorpseGUI(context.getSender(), (ServerPlayerEntity) player, deathID);
+            Guis.openCorpseGUI(context.getSender(), (ServerPlayerEntity) player, deathID);
         }
-    }
-
-    public static void openCorpseGUI(ServerPlayerEntity playerToShow, ServerPlayerEntity player, UUID uuid) {
-        Death death = DeathManager.getDeath(player, uuid);
-        if (death == null) {
-            return;
-        }
-        CorpseEntity corpse = CorpseEntity.createFromDeath(playerToShow, death);
-        NetworkHooks.openGui(playerToShow, new CorpseContainerProvider(corpse, playerToShow.abilities.isCreativeMode, true), packetBuffer -> {
-            packetBuffer.writeBoolean(true);
-            packetBuffer.writeCompoundTag(death.toNBT());
-        });
     }
 
     @Override